chore(MatrixLineagePair.java): change return type of getBlocksize to int like in the matrix object
chore(AggregateBinaryFEDInstruction.java): remove the cast to int of the blocksize

+++ src/main/java/org/apache/sysds/runtime/controlprogram/federated/MatrixLineagePair.java
@@ -85,7 +85,7 @@ public class MatrixLineagePair extends MutablePair<MatrixObject, LineageItem> {
 		return left.getNumColumns();
 	}
 
-	public long getBlocksize() {
+	public int getBlocksize() {
 		return left.getBlocksize();
 	}
